+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 1.6.1
 
 * Replace `compare_exchange` with `compare_exchange_weak` where spurious failures do not cost much.
+* Fix an issue with `hash_map::Reserve::fmt` which printed superfluous information.
 
 1.6.0